EVP: Make it possible to have keymgmt name that differs from the algo
In some cases, the key has a different type name than the algorithms
that use them.  For example, both ECDH and ECDSA uses EC keys.

To accomodate this, we give the algorithms the possibility to give
back the type name they expect for their keys.

Since the resulting keymgmt MUST be from the same provider, it's up to
the provider authors to make sure the names match.