Move the definition of OPENSSL_BUILDING_OPENSSL
It was placed in a spot that works as fallback, and would therefore
not at all for config targets that define library macros of their own.

Furthermore, this should be attached specifically to our public
libraries, and that's done more precisely in our build.info files.

We ddo not attach it to anything else, so that our programs and engine
become and therefore are closer to what our users are exposed to...
in the spirit of eating our own dog food.