PROV: make DSA keymgmt less dependent
With ex_data support in DSA keys gone for the FIPS module, there's no
need to create a key with a library context as input, i.e. we can go
back to normal DSA_new().

Also, we avoid using DSAparams_dup(), as that does a lot of ASN.1
magic that we don't actually need.  Just duplicate P, G and Q.