fixup! EVP: Adapt EVP_PKEY checking, comparing and copying for provider keys
This helps in some cases where it's not terribly important if the
provider data is a set of domain parameters or a keypair.  -1 is used
by EVP_PKEY_copy_parameters(), which allows copying from provider data
that's a keypair just as well as domain parameters.

This means that providers can't really have a strict separation
between keypairs and domain parameters, as the former is increasingly
assumed to contain the latter.