Adapt tests for SM2 changes.
Since the SM2 curve now implies SM2 computation by default,
test/ecdsatest.c is faulty for not passing an SM2 id.  We could handle
it by creating the appropriate EVP_PKEY_CTX and giving it the SM2 id,
but we do want a test of switching a key with the SM2 curve to become
a "normal" EC key with EVP_PKEY_set_alias_type(), so that's what we do
here.

test/evp_test.c, on the other hand, doesn't need to explicitly set the
EVP_PKEY_SM2 alias type, as that now happens automatically.