Allow keys without available public key to be used for SSL
Fixes: 1532

I had previously made it stop segfaulting, which was a good start.
This makes it *work*.

It's a special case for EC for now; we can later extend the generic
EVP_PKEY API to have a generic "here's a public key I checked matches
this private key; you might want to remember it" method.

But first we need to deal with the fact that X509_check_private_key(),
where we want to do this thing, now takes a 'const' private key, so it's
not clear if we should be doing *anything* of that form at all.