RAND_DRBG: Restore the unused EX_DATA field in the FIPS provider
Having structures in the FIPS and the default provider with the
same name but different memory layouts is possible in theory,
since the objects are not shared between those two providers.

However, it can have very confusing side effects in the debugger,
because the members can slip out of place and the debugger will
show misleading information.

For that reason, it is better to keep the EX_DATA field even though
it is not used in the FIPS provider.