Configuration: Add post module building check
We want an extra check that the FIPS provider module has resolved all
symbols except the standard library ones.  This is generalized with
the addition of a 'dso_post_cmd' config target attribute, that holds
the command line to do the check.

For the moment, we only do this for module building with the
'linux-shared' shared library scheme and its derivatives.

Fixes #11020