Run apps/progs.pl from the build tree
`apps/progs.pl` is optimized to be run from the build tree, but `make
update` currently runs it from the src tree.

This can lead to failures in out-of-tree builds if the build tree is
nested deeper than the src tree compared to the common ancestor
node.

This commit fixes the Makefile template to run `apps/progs.pl` from the
build tree instead.